Add the project screen: tasks and notes (F3)
Tasks and notes for one project, reachable from the board. Tasks: inline
compose, inline title edit, status, description, delete, and reorder via
ReorderableListView mapped onto PATCH /tasks/:id/position with the
neighbours taken from the loaded list. Notes: list, full-screen editor
with an edit/preview toggle and an unsaved-changes guard, rendered with
flutter_markdown_plus.

Writes are optimistic with rollback, and serialised through a one-deep
queue: without that, a second create captures the first one's prediction
as its rollback target, which is reachable just by typing two tasks
quickly.

Any mutation that can move the "first pending task" -- create, status,
delete, reorder -- re-reads that project's tasks. That is the only way to
both honour "the server computes isCurrent" and avoid holding stale
positions after a rebalance, since the position endpoint returns the moved
row and says nothing about the rows it renumbered. Title and description
edits change neither, so they merge the response instead of paying for a
round trip. The board is then updated by splicing that list into its row,
so counters, current task, badges and the reminder queue all follow with
no extra GET /board -- and with no scheduler call anywhere in F3, which
keeps "changing the target set is the rescheduling" intact.

Routing now exists, below the session switch as F0 planned, so F4's
"tap the notification, land on the task" has somewhere to go.

Two bugs that fixtures could not have caught, both found by testing
against the real server:

- Every DELETE was broken. dio's BaseOptions.contentType put
  application/json on bodiless requests and Fastify rejects an empty
  body -- the exact failure predicted when the backend error handler was
  fixed, now visible as a 400 instead of a 500.
- Task.position was parsed as int. It is a Prisma Float: four reorders
  into one slot give 2937.5, and toInt() silently returns 2937, so the
  client's ordering key diverges from the server's with no error.

316 tests, analyze clean. The live write path is covered end to end
against a real backend and real Postgres: the tests create disposable
projects, exercise create/reorder/status/description/remindAt/delete and
the note CRUD, then archive and delete themselves.

+++ app/lib/models/task.dart
@@ -23,7 +23,29 @@ abstract class Task with _$Task {
     /// reordering (F3) sends neighbour ids (`beforeTaskId` / `afterTaskId`) and
     /// lets the server pick the new value, so nothing on the client should ever
     /// compute a position itself.
-    required int position,
+    ///
+    /// ## Why `double` and not `int`
+    ///
+    /// `position` is a **Float** in the Prisma schema, and that is load-bearing:
+    /// `backend/src/domain/position.ts` picks a new position by bisecting the
+    /// gap between the two neighbours, so the fourth or fifth reorder into the
+    /// same spot produces 1062.5 and JSON carries it as `1062.5`.
+    ///
+    /// Declared `int`, this does **not** crash, which is worse than if it did:
+    /// `json_serializable` emits `(json['position'] as num).toInt()`, so 1062.5
+    /// silently becomes 1062. Two rows bisected into the same integer gap then
+    /// collapse to the same position on the client while the server has them
+    /// distinct and ordered -- a divergence with no error, no log line, and no
+    /// way to notice except by watching a list come back in a different order
+    /// than it went out. Seeded data (1000/2000/3000) can never show it; a few
+    /// drags into one spot can.
+    ///
+    /// The client reads this field for diagnostics only. **List order is the
+    /// server's order** (`GET /projects/:id/tasks` sorts by `position` for us);
+    /// nothing here ever re-sorts by it, which is what makes an optimistic
+    /// reorder -- rows in the new order still carrying their old positions --
+    /// safe for the one frame before the server's answer lands.
+    required double position,
 
     /// Calendar date (stored server-side as UTC midnight) to be reminded about a
     /// `blocked` task. Only ever compare this by its `YYYY-MM-DD` prefix -- see
